pyproject.toml
[project]
name = "openai-agent-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents[voice]>=0.2.5",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "textual>=5.3.0",
]
